ref: refactoring component Todos
