+++ src/features/ui/index.ts
@@ -5,3 +5,5 @@ export * from './edit-title'
 export * from './filter-tasks'
 export * from './tasks-list'
 export * from './sign-in-form'
+export * from './pagination'
+export * from './todos-list'
