+++ src/features/ui/pagination/index.ts
@@ -1,0 +1,1 @@
+export { TodosPagination } from './pagination'
