+++ src/features/ui/pagination/pagination.module.scss
@@ -1,0 +1,9 @@
+.pagination{
+  display: flex;
+  justify-content: center;
+
+  width: 100%;
+  max-width: 300px;
+  margin: 0 auto;
+  padding: 30px 0 20px;
+}
