+++ src/features/ui/todos-list/index.ts
@@ -1,0 +1,1 @@
+export { TodosList } from './todos-list'
