+++ src/pages/todos/todos.module.scss
@@ -39,13 +39,3 @@
 .form > button {
   width: 15%;
 }
-
-.pagination{
-  display: flex;
-  justify-content: center;
-
-  width: 100%;
-  max-width: 300px;
-  margin: 0 auto;
-  padding: 30px 0 20px;
-}
