+++ src/pages/todos/todos.tsx
@@ -1,9 +1,9 @@
 import React, { ChangeEvent, useEffect, useState } from 'react'
 
-import { AddItemForm, Todo } from '@/features'
+import { AddItemForm, TodosList, TodosPagination } from '@/features'
 import { useAppDispatch, useAppSelector } from '@/hooks'
-import { changePage, createTodo, fetchTodos, me, removeTodo } from '@/redux'
-import { CircularProgress, Grid, Pagination, ThemeProvider } from '@mui/material'
+import { createTodo, fetchTodos, me } from '@/redux'
+import { CircularProgress, ThemeProvider } from '@mui/material'
 
 import s from './todos.module.scss'
 
@@ -13,16 +13,8 @@ export const TodosPage = () => {
   const [inputValue, setInputValue] = useState<string>('')
   const [error, setError] = useState<boolean>(false)
   const { currentPage, portion, todos } = useAppSelector(state => state.todoLists)
-  const tasks = useAppSelector(state => state.tasksList)
   const isLoading = useAppSelector(state => state.auth.isLoading)
   const dispatch = useAppDispatch()
-  const pagesCount = Math.ceil(todos.length / portion)
-
-  const todosPaginated = todos.slice((currentPage - 1) * portion, currentPage * portion)
-
-  const changeCurrentPage = (_: any, page: number) => {
-    dispatch(changePage(page))
-  }
 
   const onInputChangeValue = (e: ChangeEvent<HTMLInputElement>) => {
     setInputValue(e.currentTarget.value)
@@ -42,17 +34,10 @@ export const TodosPage = () => {
       setInputValue('')
     }
   }
-  const removeTodoList = async (id: string) => {
-    await dispatch(removeTodo(id))
-  }
-
-  const getAllTasks = async () => {
-    await dispatch(fetchTodos())
-  }
 
   useEffect(() => {
-    getAllTasks()
-  }, [])
+    dispatch(fetchTodos())
+  }, [dispatch])
 
   useEffect(() => {
     dispatch(me())
@@ -81,32 +66,12 @@ export const TodosPage = () => {
               {!todos.length ? (
                 <div> You do not have any todos yet! Add your first TodoList</div>
               ) : (
-                <Grid container>
-                  {todosPaginated.map(todo => (
-                    <Grid item key={todo.id}>
-                      <Todo
-                        id={todo.id}
-                        removeTodo={removeTodoList}
-                        tasks={tasks[todo.id]}
-                        title={todo.title}
-                      />
-                    </Grid>
-                  ))}
-                </Grid>
+                <TodosList currentPage={currentPage} portion={portion} todos={todos} />
               )}
             </ThemeProvider>
           )}
         </div>
-        {todos.length > 4 && (
-          <div className={s.pagination}>
-            <Pagination
-              color={'secondary'}
-              count={pagesCount}
-              onChange={changeCurrentPage}
-              page={currentPage}
-            />
-          </div>
-        )}
+        {todos.length > 4 && <TodosPagination />}
       </div>
     </div>
   )
